uboot: fix compile warning
Summary: fix compile warning

Test Plan: no warning after build

Reviewers: huiliang.xu, dinggao.pan

Reviewed By: huiliang.xu, dinggao.pan

Subscribers: huiliang.xu

Differential Revision: https://cr.hobot.cc/D45953

+++ drivers/mmc/hobot_dw_mmc.c
@@ -149,7 +149,7 @@ static void sdio_pin_mux_config(int dev_index)
 #endif
 #endif
 #ifdef CONFIG_TARGET_X3
-	unsigned int i;
+	u64 i;
 	if (dev_index == 0) {
 		for (i = SD0_CLK; i <= SD0_WPROT; i += 4) {
 			reg_val = readl(i);
